yeah, no idea what I did :3

+++ Zombie Server of Doom/wServer/realm/terrain/Tile.cs
@@ -1,7 +1,7 @@
 ﻿
 namespace wServer.realm.terrain
 {
-    public enum Tile : byte
+    public enum Tile : ushort
     {
     }
 }
